Fix handling of csrNonce in MTRCommissioningParameters. (#42075)
* Fix handling of csrNonce in MTRCommissioningParameters.

The CSR nonce passed via CommissioningParameters was being ignored by
AutoCommisssioner, because it was just fetching a random nonce from the
OperationalCredentialsDelegate.

This fix makes it possible for an OperationalCredentialsDelegate to indicate
that the AutoCommissioner should use whatever nonce was provided via
CommissioningParameters, and uses that in MTROperationalCredentialsDelegate, so
we don't override the data from MTRCommissioningParameters.

* Apply suggestion from code review.

Co-authored-by: Karsten Sperling <113487422+ksperling-apple@users.noreply.github.com>

* Address review comment.

---------

Co-authored-by: Karsten Sperling <113487422+ksperling-apple@users.noreply.github.com>