Python framework: Remove default from get_endpoint (#41711)
* Python framework: Remove default from get_endpoint

In order to do test selection we need to know the endpoint
before the test. if the default is set by the test explicitly
outside of the test function call, we know.
Ideally, all tests would be explicit about asking the tester
to set the endpoint to avoid accidentally testing the wrong
thing. However, the intent of this PR is not to change that
policy, but rather to make the information available before
the test function itself is run, either through the user
command line parameter or through the static default.

The functions inside tests were adjusted according to the following
rules
- get_endpoint calls in test_ functions with @run_if_endpoint_matches
  decorators just have the defaults removed. Those decorators
  already enforce that the endpoint parameter is given on the
  command line so the default would never be used.
- get_endpoint calls in utility functions have the defaults
  removed since these should use the endpoint of the caller
  in the first place
- get_endpoint calls with a default in functions decorated with
  the async_test_body decorator have an explicit default_endpoint
  property function defined to match the given default if the
  default is not 0

* Fix type annotation

* A couple more

* Fix doorlock

* fix webrtcp-2.9

* fix a couple more

* Fix two additional tests added after this PR was raised

* missed one more

* add default endpoints to door lock

* I liked the function so much I thought...why not two?